pyproject.toml
[project]
name = "nova-health-bot"
version = "0.1.0"
description = "N.O.V.A. MVP Comm Link"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.19",
    "langchain-openai>=1.1.10",
    "langgraph>=0.2.74",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.11.1",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.0",
    "typer>=0.15.1",
    "gspread>=6.0.0",
    "google-auth>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.5",
    "ruff>=0.9.9",
]

[project.scripts]
nova = "nova.cli:app"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
strict = true
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nova"]
